basic rotate tests

+++ tests/regression/tools/rotation/rotate_utils.sh
@@ -2,8 +2,8 @@ function validate_test_chunks ()
 {
 	TRACE_PATH=$1
 	today=$2
-	domain=$3
 	app_path=$3
+	domain=$4
 
 	# YYYYMMDD-HHMMSS-YYYYMMDD-HHMMSS
 	chunk_pattern="${today}-[0-9][0-9][0-9][0-9][0-9][0-9]-${today}-[0-9][0-9][0-9][0-9][0-9][0-9]"
@@ -14,14 +14,13 @@ function validate_test_chunks ()
 	ls $TRACE_PATH/${chunk_pattern}-2/${app_path}/metadata >/dev/null
 	ok $? "Chunk 2 exists"
 	# FIXME
-#	ls $TRACE_PATH/${chunk_pattern}-3/${app_path}/metadata >/dev/null
-#	ok $? "Chunk 3 exists"
+	ls $TRACE_PATH/${chunk_pattern}-3/${app_path}/metadata >/dev/null
+	ok $? "Chunk 3 exists"
 
 	# Make sure we don't have anything else in the first 2 chunk directories
 	# besides the kernel folder.
 	nr_stale=$(\ls $TRACE_PATH/${chunk_pattern}-1 | grep -v $domain | wc -l)
 	ok $nr_stale "No stale folders in chunk 1 directory"
-	\ls $TRACE_PATH/${chunk_pattern}-1
 	nr_stale=$(\ls $TRACE_PATH/${chunk_pattern}-2 | grep -v $domain | wc -l)
 	ok $nr_stale "No stale folders in chunk 2 directory"
 
@@ -43,20 +42,19 @@ function validate_test_chunks ()
 	fi
 
 	# Chunk 3: 0 event
-	# FIXME
-#	validate_trace_empty $TRACE_PATH/${chunk_pattern}-3
-#	if [ $? -eq 0 ]; then
-#		# Only delete if successful
-#		rm -rf $TRACE_PATH/${chunk_pattern}-3
-#	fi
+	validate_trace_empty $TRACE_PATH/${chunk_pattern}-3
+	if [ $? -eq 0 ]; then
+		# Only delete if successful
+		rm -rf $TRACE_PATH/${chunk_pattern}-3
+	fi
 
 	# The session folder after all chunks have been removed is empty
-#	test -z "$(\ls -A $TRACE_PATH)"
-#	empty=$?
-#	ok $empty "Trace folder is now empty"
-#	if [ $empty -eq 0 ]; then
-		# Only delete if successful
+	test -z "$(\ls -A $TRACE_PATH)"
+	empty=$?
+	ok $empty "Trace folder is now empty"
+	if [ $empty -eq 0 ]; then
+   	# Only delete if successful
 		rm -rf $TRACE_PATH/
-#	fi
+	fi
 }
